feat mapper代理兼容增删改返回类型

+++ src/main/java/com/jojo/demo/mapper/UserMapper.java
@@ -20,10 +20,10 @@ public interface UserMapper {
     User selectOne(@Param("id")Integer id);
 
     @Insert("insert into t_user(name, age) values(#{user.name}, #{user.age})")
-    Integer insert(@Param("user") User user);
+    Long insert(@Param("user") User user);
 
     @Delete("delete from t_user where id = #{id}")
-    Integer delete(@Param("id")Integer id);
+    Long delete(@Param("id")Integer id);
 
     @Update("update t_user set name = #{name} where id = #{id}")
     Integer update(@Param("id")Integer id, @Param("name")String name);
